Lab 3: Created library for the LCD used in lab
Part 1 had no code to write, hence no code to commit. There might be something
for part 2, namely use bluetooth to communicate textual data.

Wrote a library for the LCD used in lab, to have a clean interface when working
with it. This allows us to have a object which we can call functions on, e.g.
  pinMode(...)  => Display lcd(...)
  char2LCD('a') => lcd.charWrite('a')
  str2LCD("ab") => lcd.strWrite("ab")

See `the libnhdlcd/example` for more.